+++ parser_library/src/lsp/item_convertors.h
@@ -16,6 +16,7 @@
 #define HLASMPLUGIN_PARSERLIBRARY_LSP_ITEM_CONVERTORS_H
 
 #include <functional>
+#include <span>
 #include <string>
 #include <string_view>
 #include <vector>
@@ -27,6 +28,7 @@ namespace context {
 class macro_definition;
 struct sequence_symbol;
 class symbol;
+struct using_context_description;
 } // namespace context
 } // namespace hlasm_plugin::parser_library
 
@@ -40,6 +42,8 @@ struct variable_symbol_definition;
 
 std::string hover_text(const context::symbol& sym);
 std::string hover_text(const variable_symbol_definition& sym);
+std::string hover_text(std::span<const context::using_context_description> usings);
+void append_hover_text(std::string& buffer, const context::using_context_description& u);
 std::string get_macro_documentation(const text_data_view& macro_text, size_t definition_line);
 std::string get_macro_signature(const context::macro_definition& m);
 bool is_continued_line(std::string_view line);