+++ parser_library/src/lsp/macro_info.h
@@ -15,10 +15,15 @@
 #ifndef LSP_MACRO_INFO_H
 #define LSP_MACRO_INFO_H
 
-#include <map>
+#include <vector>
 
 #include "context/macro.h"
 #include "symbol_occurrence.h"
+#include "tagged_index.h"
+
+namespace hlasm_plugin::parser_library::context {
+class using_collection;
+} // namespace hlasm_plugin::parser_library::context
 
 namespace hlasm_plugin::parser_library::lsp {
 
@@ -100,8 +105,16 @@ struct macro_slice_t
 using file_scopes_t = std::unordered_map<utils::resource::resource_location,
     std::vector<lsp::macro_slice_t>,
     utils::resource::resource_location_hasher>;
+
+struct line_occurence_details
+{
+    size_t max_endline = 0;
+    index_t<context::using_collection> active_using;
+    bool using_overflow = false;
+};
+
 using file_occurrences_t = std::unordered_map<utils::resource::resource_location,
-    std::pair<std::vector<symbol_occurrence>, std::map<size_t, size_t>>,
+    std::pair<std::vector<symbol_occurrence>, std::vector<line_occurence_details>>,
     utils::resource::resource_location_hasher>;
 
 class lsp_context;