+++ parser_library/src/processing/statement_analyzers/lsp_analyzer.cpp
@@ -137,15 +137,37 @@ const expressions::mach_expr_symbol* get_single_mach_symbol(const semantics::ope
 
 void lsp_analyzer::analyze(const semantics::preprocessor_statement_si& statement)
 {
-    auto collection_info = get_active_collection(hlasm_ctx_.opencode_location(), false);
+    auto ci = get_active_collection(hlasm_ctx_.opencode_location(), false);
 
-    auto& end_line = (*collection_info.stmt_ranges)[statement.m_details.stmt_r.start.line];
-    end_line = std::max(end_line, statement.m_details.stmt_r.end.line);
+    collect_endline(statement.m_details.stmt_r, ci);
 
-    collect_occurrences(lsp::occurrence_kind::ORD, statement, collection_info);
+    collect_occurrences(lsp::occurrence_kind::ORD, statement, ci);
 
     if (const auto& operands = statement.m_details.operands; statement.m_copylike && operands.size() == 1)
-        add_copy_operand(hlasm_ctx_.ids().add(operands.front().name), operands.front().r, collection_info);
+        add_copy_operand(hlasm_ctx_.ids().add(operands.front().name), operands.front().r, ci);
+}
+
+
+lsp::line_occurence_details& lsp_analyzer::line_details(const range& r, const collection_info_t& ci)
+{
+    if (r.start.line >= ci.line_details->size())
+        ci.line_details->insert(ci.line_details->end(), r.start.line + 1 - ci.line_details->size(), {});
+    return (*ci.line_details)[r.start.line];
+}
+
+void lsp_analyzer::collect_endline(const range& r, const collection_info_t& ci)
+{
+    auto& line_detail = line_details(r, ci);
+    line_detail.max_endline = std::max(line_detail.max_endline, r.end.line + 1);
+}
+
+void lsp_analyzer::collect_usings(const range& r, const collection_info_t& ci)
+{
+    auto& line_detail = line_details(r, ci);
+    if (auto cur = hlasm_ctx_.using_current(); !line_detail.active_using)
+        line_detail.active_using = cur;
+    else
+        line_detail.using_overflow |= line_detail.active_using != cur;
 }
 
 void lsp_analyzer::macrodef_started(const macrodef_start_data& data)
@@ -203,18 +225,26 @@ void lsp_analyzer::collect_occurrences(
 
     if (auto def_stmt = statement.access_deferred())
     {
-        const auto& stmt_range = def_stmt->stmt_range_ref();
-        auto& end_line = (*ci.stmt_ranges)[stmt_range.start.line];
-        end_line = std::max(end_line, stmt_range.end.line);
+        if (def_stmt->instruction_ref().type != semantics::instruction_si_type::EMPTY)
+        {
+            const auto& r = def_stmt->stmt_range_ref();
+            collect_endline(r, ci);
+            collect_usings(r, ci);
+        }
+
         collect_occurrence(def_stmt->label_ref(), collector);
         collect_occurrence(def_stmt->instruction_ref(), collector);
         collect_occurrence(def_stmt->deferred_ref(), collector);
     }
     else if (auto res_stmt = statement.access_resolved())
     {
-        const auto& stmt_range = res_stmt->stmt_range_ref();
-        auto& end_line = (*ci.stmt_ranges)[stmt_range.start.line];
-        end_line = std::max(end_line, stmt_range.end.line);
+        if (res_stmt->instruction_ref().type != semantics::instruction_si_type::EMPTY)
+        {
+            const auto& r = res_stmt->stmt_range_ref();
+            collect_endline(r, ci);
+            collect_usings(r, ci);
+        }
+
         collect_occurrence(res_stmt->label_ref(), collector);
         collect_occurrence(res_stmt->instruction_ref(), collector);
         collect_occurrence(res_stmt->operands_ref(), collector);